Convert StripLocation tests to jtreg
